refactor functions in product-entry.js into entry-utils.js and update local storage with data
