+++ constants.js
@@ -1,1 +1,2 @@
 export const CART = 'cart';
+export const PRODUCTS = 'products';
